Crate futures_micro

A very small, no-std compatible toolbox of async utilities.

Re-exports

pub use core::future::Future;

pub use core::pin::Pin;

pub use core::task::Context;

pub use core::task::Poll;

pub use core::task::Waker;

Modules

prelude

This module has extras that clash with names in [futures-lite], which depends on us.

Macros

or

Polls arbitrarily many futures, returning the first ready value.

pin

Pins a variable of type T on the stack and rebinds it as Pin<&mut T>.

ready

Unwraps Poll<T> or returns Pending.

zip

Zips arbitrarily many futures, waiting for all to complete.

Structs

Or

Returns the result of left or right future, preferring left if both are ready.

PollFn

Future for the poll_fn() function.

Zip

Waits for two Futures to complete, returning both results.

Functions

poll_fn

Creates a future from a function returning Poll.

sleep

Goes to sleep until woken by its Waker being called.

waker

Get the Waker inside an async fn where you aren’t supposed to have it.

yield_once

Pushes itself to the back of the executor queue so some other tasks can do some work.